// types/player.ts
export interface Player {
  idPlayer: string;
  strPlayer: string;
  strPlayerAlternate?: string;
  strNationality?: string;
  strTeam?: string;
  idTeam?: string;
  strSport?: string;
  dateBorn?: string;
  strBirthLocation?: string;
  strDescriptionEN?: string;
  strPosition?: string;
  strHeight?: string;
  strWeight?: string;
  strThumb?: string;
  strCutout?: string;
  strRender?: string;
  strBanner?: string;
  strFanart1?: string;
  strNumber?: string;
  strWage?: string;
  strSigning?: string;
  strStatus?: string;
  strGender?: string;
  strSide?: string;
}